Dedupe small Expr helpers onto canonical implementations
- contains_inexact_real / is_inexact / contains_real /
  contains_inexact_literal (9 copies) -> math_ast::contains_inexact_real
- expr_contains_real (2), is_concrete_number (2) -> math_ast
- expr_to_rational / extract_rational (5 copies) -> math_ast::expr_to_rational
- expr_to_f64 (3 copies) -> math_ast::expr_to_f64
- bool_expr (5 copies) -> syntax::bool_expr
- format_bigfloat_value, is_prime_i128, cmul, assoc_get, snap_coord,
  make_quantity, expr_str_eq, nested is_pi: single canonical each
- evaluator::type_helpers now re-exports num_to_expr, expr_to_i128,
  expr_to_bigint, constant_to_f64, needs_bigint from math_ast instead of
  duplicating them
