+++ src/functions/expr_form.rs
@@ -501,18 +501,7 @@ fn expr_contains_imag_unit(expr: &Expr) -> bool {
 /// together with `expr_contains_imag_unit` to decide whether a sum
 /// like `Real + Real*I` should collapse to `Complex[re\`, im\`]` in
 /// FullForm.
-fn expr_contains_real(expr: &Expr) -> bool {
-  match expr {
-    Expr::Real(_) => true,
-    Expr::List(items) => items.iter().any(expr_contains_real),
-    Expr::FunctionCall { args, .. } => args.iter().any(expr_contains_real),
-    Expr::BinaryOp { left, right, .. } => {
-      expr_contains_real(left) || expr_contains_real(right)
-    }
-    Expr::UnaryOp { operand, .. } => expr_contains_real(operand),
-    _ => false,
-  }
-}
+use crate::functions::math_ast::expr_contains_real;
 
 /// Render a machine-precision Real for FullForm/InputForm output:
 /// integer values lose their trailing zero (`1.0` → `1.`) and every
