+++ src/functions/math_ast/gamma.rs
@@ -484,17 +484,6 @@ fn gamma_half_expr(numer: BigInt, denom: BigInt, is_neg: bool) -> Expr {
 /// number anywhere in the tree. Used to distinguish exact symbolic
 /// arguments (which Gamma should not evaluate numerically) from inexact
 /// ones (which should be evaluated to floats).
-fn contains_inexact_real(expr: &Expr) -> bool {
-  match expr {
-    Expr::Real(_) | Expr::BigFloat(_, _) => true,
-    Expr::UnaryOp { operand, .. } => contains_inexact_real(operand),
-    Expr::BinaryOp { left, right, .. } => {
-      contains_inexact_real(left) || contains_inexact_real(right)
-    }
-    Expr::FunctionCall { args, .. } => args.iter().any(contains_inexact_real),
-    _ => false,
-  }
-}
 
 /// Upper incomplete gamma function Gamma[a, z]
 fn gamma_incomplete_upper(
