+++ src/functions/quantity_ast.rs
@@ -2715,23 +2715,7 @@ pub fn try_quantity_divide(
   }
 }
 
-/// Extract a rational number (p, q) from an expression.
-/// Integer(n) → (n, 1), Rational[p, q] → (p, q).
-fn expr_to_rational(expr: &Expr) -> Option<(i128, i128)> {
-  match expr {
-    Expr::Integer(n) => Some((*n, 1)),
-    Expr::FunctionCall { name, args }
-      if name == "Rational" && args.len() == 2 =>
-    {
-      if let (Expr::Integer(p), Expr::Integer(q)) = (&args[0], &args[1]) {
-        Some((*p, *q))
-      } else {
-        None
-      }
-    }
-    _ => None,
-  }
-}
+use crate::functions::math_ast::expr_to_rational;
 
 /// Raise a unit expression to a rational power (p/q) by decomposing the unit
 /// into its base components, multiplying each exponent by p/q, and rebuilding.
