+++ src/syntax.rs
@@ -14034,6 +14034,11 @@ pub fn replace_identifier_in_expr(
   }
 }
 
+/// Build the boolean symbol `True` or `False`.
+pub fn bool_expr(b: bool) -> Expr {
+  Expr::Identifier(if b { "True" } else { "False" }.to_string())
+}
+
 pub fn unevaluated(name: &str, args: &[Expr]) -> Expr {
   Expr::FunctionCall {
     name: name.to_string(),
